Extract title of press releases from Department of Health website

Extract title of press releases from Department of Health website

get_pr_url(base = "https://www.doh.gov.ph/press-releases", pages = 1:25)

Arguments

base
Base URL for press releases in the Department of Health website. Default is https://www.doh.gov.ph/press-releases

pages
A vector of page numbers corresponding to the page panel containing the press release link. Default is 1:25. Press releases only go up to page 25. Values higher than 25 will error.

Value

A tibble of 2 columns: 1) press release title; 2) press release unique identifier; and, 3) date of press release.

Examples

## Get press release URLs for first page of press release panel. The base
## argument has been specified using non-SSL version of base URL to prevent
## erroring for those with proxy server connections (i.e., in Travis tests)
get_pr_url(base = "http://www.doh.gov.ph/press-releases",
           pages = 1)

#> # A tibble: 15 x 3
#>    url                                                             id date      
#>    <chr>                                                        <dbl> <date>    
#>  1 /doh-press-release/NEW-QUARANTINE-SITE-IN-ALABANG-INAUGURAT…  1452 2020-05-13
#>  2 /doh-press-release/STATEMENT-ON-DATA-INTEGRITY                1920 2020-05-13
#>  3 /doh-press-release/HONORING-NURSE%E2%80%99S-DAY%3A-DOH-CALL…  8151 2020-05-13
#>  4 /doh-press-release/SAN-JUAN-CITY-THANKS-DOH-AS-COVID-19-CAS…  1911 2020-05-13
#>  5 /press-release/ADB-Sponsored-COVID-19-Lab-in-Pampanga-Launc…  1421 2020-05-10
#>  6 /press-release/ECQ-Buys-PH-Time-Continued-Practice-of-Healt…  5317 2020-05-09
#>  7 /press-release/biggest-mega-swabbing-center-in-moa-arena-to…  2977 2020-05-08
#>  8 /doh-press-release/Press%20Release/DUQUE-THANKS-NAVY-FRONTL…  4211 2020-05-07
#>  9 /doh-press-release/PH-GOV%E2%80%99T-RECEIVES-7-METRIC-TONS-…  2018 2020-05-07
#> 10 /doh-press-release/BEYOND-NUMBERS%3A-WHAT-THE-FLATTENING-CU…  2525 2020-05-07
#> 11 /doh-press-release/2ND-MEGA-SWABBING-CENTER-SET-TO-OPERATE%…  1441 2020-05-06
#> 12 /doh-press-release/NEW-MEGA-SWABBING-CENTER-TO-RAMP-UP-COVI…  1452 2020-05-06
#> 13 /doh-press-release/DOH%3A-NO-NEW-CASES-IN-41-PROVINCES%3B-P…  1141 2020-05-06
#> 14 /doh-press-release/PH-STARTS-TO-FLATTEN-CURVE%3B-ECQ%2C-PHY…  1681 2020-05-06
#> 15 /doh-press-release/DOH-LAUNCHES-MOBILE-PHONE-SURVEY-PART-2-…  4158 2020-05-05